MonoSLAM/pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MonoSLAM"
version = "0.1.0"

# Minimum Python version
requires-python = ">=3.10"

# Direct runtime dependencies
dependencies = [
    "numpy",
    "Pillow",
    "PyYAML",
]

[dependency-groups]
dev = [
    "pytest",
]

[tool.setuptools]
# Use the common src/ layout
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
# Discover packages under src/
where = ["src"]
